versión casi estable

+++ pyQT version/psimplerenamer-qt.py
@@ -25,9 +25,7 @@ class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
         MainWindow.setObjectName("MainWindow")
         MainWindow.setWindowFlags(Qt.WindowMaximizeButtonHint|Qt.WindowCloseButtonHint)
-        MainWindow.setFixedSize(580,400)
-        
-        
+        MainWindow.setFixedSize(750,400)
         
         self.centralwidget = QtWidgets.QWidget(MainWindow)
         self.centralwidget.setObjectName("centralwidget")
@@ -41,22 +39,23 @@ class Ui_MainWindow(object):
         self.listaArchivos = QtWidgets.QTableWidget(self.centralwidget)
         self.listaArchivos.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
         #self.listaArchivos.setGeometry(QtCore.QRect(30, 60, 500, 191))
-        self.listaArchivos.setMaximumSize(QtCore.QSize(500, 16777215))
+        self.listaArchivos.setMaximumSize(QtCore.QSize(700, 16777215))
         self.listaArchivos.setSelectionBehavior(QtWidgets.QTableView.SelectRows)
-        self.listaArchivos.setSelectionMode(QtWidgets.QAbstractItemView.SingleSelection)
+        #self.listaArchivos.setSelectionMode(QtWidgets.QAbstractItemView.SingleSelection)
         self.listaArchivos.setGridStyle(QtCore.Qt.DotLine)
         self.listaArchivos.setWordWrap(True)
         self.listaArchivos.setCornerButtonEnabled(True)
         #self.listaArchivos.setRowCount(7)
         self.listaArchivos.setColumnCount(3)
         self.listaArchivos.setObjectName("listaArchivos")
+        self.listaArchivos.setMouseTracking(True)
         
-        # Columna Ruta        
+       
+        # Columna Ruta Resumida
         item = QtWidgets.QTableWidgetItem()
         self.listaArchivos.setHorizontalHeaderItem(0, item)
         self.listaArchivos.setColumnWidth(0,110)
-        #self.listaArchivos.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-             
+        
         # Columna Archivo
         item = QtWidgets.QTableWidgetItem()
         self.listaArchivos.setHorizontalHeaderItem(1, item)
@@ -70,33 +69,47 @@ class Ui_MainWindow(object):
         self.listaArchivos.horizontalHeader().setSectionResizeMode(2, QHeaderView.Stretch)
         self.listaArchivos.verticalHeader().setVisible(False)
 
+        # Columna Ruta        
+        # item = QtWidgets.QTableWidgetItem()
+        # self.listaArchivos.setHorizontalHeaderItem(3, item)
+        # self.listaArchivos.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
+
        # Botón Agregar Archivos        
         self.btnPushFile = QtWidgets.QPushButton(self.gridLayoutWidget)
         #self.btnPushFile.setGeometry(QtCore.QRect(550, 60, 31, 34))
-        self.btnPushFile.setObjectName("btnPushFile")        
-        self.btnPushFile.setText("+")
+        self.btnPushFile.setObjectName("btnPushFile")
         self.btnPushFile.setMaximumWidth(34)
+        self.btnPushFile.setIcon(QtGui.QIcon("icons/list-add.svg"))
+        self.btnPushFile.setToolTip("Agregar archivos") 
 
         # Boton Quitar Archivos        
         self.btnPullFile = QtWidgets.QPushButton(self.centralwidget)
         #self.btnPullFile.setGeometry(QtCore.QRect(550, 100, 31, 34))
         self.btnPullFile.setObjectName("btnPullFile")
-        self.btnPullFile.setText("-")
         self.btnPullFile.setMaximumWidth(34)
-                
+        self.btnPullFile.setIcon(QtGui.QIcon("icons/list-remove.svg"))
+        self.btnPullFile.setToolTip("Quitar el/los archivo(s) seleccionado(s)")
         
         # Botón Subir
         self.btnMoveUp = QtWidgets.QPushButton(self.centralwidget)
         self.btnMoveUp.setObjectName("btnMoveUp")
-        self.btnMoveUp.setText("S")
-        self.btnMoveUp.setMaximumWidth(34)        
+        self.btnMoveUp.setMaximumWidth(34)    
+        self.btnMoveUp.setIcon(QtGui.QIcon("icons/go-up.svg"))
+        self.btnMoveUp.setToolTip("Subir el/los archivo(s) seleccionado(s)")
         
         # Botón Bajar
         self.btnMoveDown = QtWidgets.QPushButton(self.centralwidget)
         self.btnMoveDown.setObjectName("btnMoveDown")
-        self.btnMoveDown.setText("B")
         self.btnMoveDown.setMaximumWidth(34)        
-        
+        self.btnMoveDown.setIcon(QtGui.QIcon("icons/go-down.svg"))
+        self.btnMoveDown.setToolTip("Bajar el/los archivo(s) seleccionado(s)")        
+
+        # Botón Limpiar
+        self.btnClean = QtWidgets.QPushButton(self.centralwidget)
+        self.btnClean.setObjectName("btnClean")
+        self.btnClean.setMaximumWidth(34)        
+        self.btnClean.setIcon(QtGui.QIcon("icons/edit-delete.svg"))
+        self.btnClean.setToolTip("Eliminar selección")     
         
         # Nuevo Nombre
         self.nuevoNombre = QtWidgets.QLineEdit(self.centralwidget)
@@ -116,6 +129,7 @@ class Ui_MainWindow(object):
         self.spinSequence = QtWidgets.QSpinBox(self.centralwidget)
         #self.spinSequence.setGeometry(QtCore.QRect(150, 300, 51, 32))
         self.spinSequence.setMinimum(1)
+        self.spinSequence.setMaximum(100)
         self.spinSequence.setObjectName("spinSequence")
         self.spinSequence.setToolTip("La secuencia comenzará a partir del número que ingrese")
 
@@ -142,12 +156,12 @@ class Ui_MainWindow(object):
 
 #================       LAYOUT       ================
 
-        # Layout arriba
-        
+        # Layout arriba izquierda
         hboxTopLeft = QHBoxLayout()
         hboxTopLeft.setObjectName("hboxTopLeft")
         hboxTopLeft.addWidget(self.listaArchivos)
         
+        # Layout arriba derecha
         vboxTopRight = QVBoxLayout()
         vboxTopRight.setAlignment(Qt.AlignTop)
         vboxTopRight.setContentsMargins(0, -1, -1, 10)
@@ -155,7 +169,9 @@ class Ui_MainWindow(object):
         vboxTopRight.addWidget(self.btnPullFile)
         vboxTopRight.addWidget(self.btnMoveUp)
         vboxTopRight.addWidget(self.btnMoveDown)
-
+        vboxTopRight.addWidget(self.btnClean)
+        
+        # Layout que contiene los dos layouts Top
         vboxFirst = QHBoxLayout()
         vboxFirst.addLayout(hboxTopLeft)
         vboxFirst.addLayout(vboxTopRight)
@@ -186,20 +202,21 @@ class Ui_MainWindow(object):
         #hboxBottomRight.setAlignment(Qt.AlignTop)
         hboxBottomRight.addWidget(self.btnRename)
         hboxBottomRight.addWidget(self.btnClose)
+  
                
         vboxRight = QVBoxLayout()
         vboxRight.addStretch(1)
         vboxRight.addLayout(hboxBottomRight)
 
     
-        self.gridLayoutWidget.setGeometry(QtCore.QRect(30, 30, 500, 250))
+        self.gridLayoutWidget.setGeometry(QtCore.QRect(30, 30, 700, 250))
         self.gridLayout1 = QtWidgets.QGridLayout(self.gridLayoutWidget)
         self.gridLayout1.setContentsMargins(0, 0, 0, 0)
         self.gridLayout1.addLayout(vboxFirst,0,0)
         #self.gridLayout1.addLayout(vboxTopRight,0,1)        
 
         self.gridLayoutWidget = QtWidgets.QWidget(self.centralwidget)
-        self.gridLayoutWidget.setGeometry(QtCore.QRect(30, 285, 500, 85))
+        self.gridLayoutWidget.setGeometry(QtCore.QRect(30, 285, 700, 85))
         self.gridLayoutWidget.setObjectName("gridLayoutWidget")
         self.gridLayout = QtWidgets.QGridLayout(self.gridLayoutWidget)
         self.gridLayout.setContentsMargins(0, 0, 0, 0)
@@ -227,15 +244,19 @@ class Ui_MainWindow(object):
         # Texto de las columnas de la tabla de Archivos
         item = self.listaArchivos.horizontalHeaderItem(0)
         item.setText(_translate("MainWindow", "Ruta"))
+        
         item = self.listaArchivos.horizontalHeaderItem(1)
         item.setText(_translate("MainWindow", "Nombre"))
         item = self.listaArchivos.horizontalHeaderItem(2)
         item.setText(_translate("MainWindow", "Nuevo Nombre"))
         
+     #   item = self.listaArchivos.horizontalHeaderItem(3)
+      #  item.setText(_translate("MainWindow", "Ruta"))
+        
+        
         # Do the resize of the columns by content
         #self.listaArchivos.resizeColumnsToContents()        
-        #self.grid_layout.addWidget(self.listaArchivos, 0, 0)   # Adding the table to the grid
-        
+        #self.grid_layout.addWidget(self.listaArchivos, 0, 0)   # Adding the table to the grid        
 
 class SimpleRenamer(QtWidgets.QMainWindow):
     def __init__(self):
@@ -244,7 +265,7 @@ class SimpleRenamer(QtWidgets.QMainWindow):
         self.ui.setupUi(self)
         
         # Eventos de botones
-        self.ui.btnClose.clicked.connect(QtWidgets.QApplication.instance().quit)
+
         self.ui.btnPushFile.clicked.connect(self.callOpenDialog)
         self.ui.btnPullFile.clicked.connect(self.removeFiles)
         self.ui.spinSequence.valueChanged.connect(self.synchNumbers)
@@ -252,8 +273,10 @@ class SimpleRenamer(QtWidgets.QMainWindow):
         self.ui.nuevoNombre.textChanged.connect(self.synchNumbers)
         self.ui.btnMoveUp.clicked.connect(self.moveUp)
         self.ui.btnMoveDown.clicked.connect(self.moveDown)
+        self.ui.btnClean.clicked.connect(self.cleanList)
         self.ui.btnRename.clicked.connect(self.applyRenaming)
-
+        self.ui.btnClose.clicked.connect(QtWidgets.QApplication.instance().quit)
+        
     # Lógica para renombrar los archivos según los parámetros elegidos en pantalla
     def renameFiles(self, inSrcFile, inSequence):
     
@@ -306,12 +329,11 @@ class SimpleRenamer(QtWidgets.QMainWindow):
             
             if not indexes:
                 return
-            
-            for index in sorted(indexes):
-                intRow = index.row()
-                print('Row %d is selected' % intRow)                
-                
-                if direction == "UP":
+                 
+            if direction == "UP":
+                for index in sorted(indexes):
+                    intRow = index.row()
+                    print('Row %d is selected' % intRow)                     
                     if intRow > 0:
                         self.ui.listaArchivos.insertRow(intRow - 1)
                         for i in range(self.ui.listaArchivos.columnCount()):
@@ -319,31 +341,49 @@ class SimpleRenamer(QtWidgets.QMainWindow):
                             #self.ui.listaArchivos.setCurrentCell(index.row()-1,column)          
                         self.ui.listaArchivos.removeRow(intRow + 1)
                         
-                else:
-                    if intRow < int(self.ui.listaArchivos.columnCount()):
+                        # Mueve el foco a la fila movida 
+                        newIndex = self.ui.listaArchivos.selectionModel().model().index(intRow - 1,0)
+                        self.ui.listaArchivos.selectionModel().select(newIndex,QItemSelectionModel.Select)
+                        self.ui.listaArchivos.setCurrentIndex(newIndex)
+            else:
+                for index in sorted(indexes, reverse=True):
+                    intRow = index.row()
+                    print('Row %d is selected' % intRow)     
+                    if intRow < int(self.ui.listaArchivos.rowCount()-1):
                         self.ui.listaArchivos.insertRow(intRow + 2)
                         for i in range(self.ui.listaArchivos.columnCount()):
-                            self.ui.listaArchivos.setItem(intRow + 2,i,self.ui.listaArchivos.takeItem(intRow,i))
+                            self.ui.listaArchivos.setItem(intRow + 2, i,self.ui.listaArchivos.takeItem(intRow,i))
                             #self.ui.listaArchivos.setCurrentCell(intRow + 2,column)
                         self.ui.listaArchivos.removeRow(intRow)
                         
-            
+                        # Mueve el foco a la fila movida 
+                        newIndex = self.ui.listaArchivos.selectionModel().model().index(intRow + 1,0)
+                        self.ui.listaArchivos.selectionModel().select(newIndex,QItemSelectionModel.Select)
+                        self.ui.listaArchivos.setCurrentIndex(newIndex)
+                
+                self.ui.listaArchivos.setFocus()
+        
     # Mover arriba
     def moveUp(self):
         self.moveFiles("UP")
         self.synchNumbers()
 
+    # Mover abajo
     def moveDown(self):
         self.moveFiles("DOWN")
         self.synchNumbers()
         
-        
+    # Quitar archivos
     def removeFiles(self):
         indexes = self.ui.listaArchivos.selectionModel().selectedRows()
         for index in sorted(indexes):
             self.ui.listaArchivos.removeRow(index.row())
         self.synchNumbers()
-        
+ 
+    # Limpia la tabla
+    def cleanList(self):
+        self.ui.listaArchivos.selectionModel().model().removeRows(0, self.ui.listaArchivos.rowCount())
+
         
     # Renombra físicamente los archivos
     def applyRenaming(self):
@@ -351,17 +391,26 @@ class SimpleRenamer(QtWidgets.QMainWindow):
         print ("Renombrado de archivos")
         # Recorre la tabla de archivos
         for nbrIndex in range(self.ui.listaArchivos.rowCount()):
-            srcFolder = self.ui.listaArchivos.item(nbrIndex,0)
+            srcFolder = self.ui.listaArchivos.item(nbrIndex,0).toolTip()
             srcFile = self.ui.listaArchivos.item(nbrIndex,1)
             dstFile = self.ui.listaArchivos.item(nbrIndex,2)
         
-            src = srcFolder.text() + srcFile.text()
-            dst = srcFolder.text() + dstFile.text()
+            src = srcFolder + srcFile.text()
+            dst = srcFolder + dstFile.text()
             
             print (src + " >>> " + dst)
 
             #os.rename(src,dst)
+        print ("Renombrado finalizado")
+        self.cleanList()
 
+    def itemAlreadyExists(self, inSrcFile):
+        
+        itemList = self.ui.listaArchivos.findItems(inSrcFile, QtCore.Qt.MatchExactly)
+        if len(itemList) > 0:
+            return True
+        else:
+            return False
 
     def callOpenDialog(self):
         self.openFileNamesDialog()
@@ -380,7 +429,7 @@ class SimpleRenamer(QtWidgets.QMainWindow):
         
         options = QFileDialog.Options()
         options |= QFileDialog.DontUseNativeDialog        
-        files, _ = QFileDialog.getOpenFileNames(self,"QFileDialog.getOpenFileNames()", "","All Files (*);;Python Files (*.py)", options=options)
+        files, _ = QFileDialog.getOpenFileNames(self,"Agregar archivos", "","All Files (*);;Python Files (*.py)", options=options)
         
         if files:
 
@@ -388,29 +437,41 @@ class SimpleRenamer(QtWidgets.QMainWindow):
 #                filename = os.path.normpath(filename)
                 srcFile = filename.split("/")[-1]             
                 srcFolder = filename.split(srcFile,1)[0]
+                lenSplit = len(srcFolder.split("/")) - 2
+                srcFolderSplit = "../" + srcFolder.split("/")[lenSplit] + "/" 
                 
+                       
 #                srcFolder = filename.directory() + "/"
 #                srcFile = filename.split(srcFolder,1)            
                 # debug print ("File " + srcFile)
                 # debug print ("Path " + srcFolder)
                 
-                itemFolder = QtWidgets.QTableWidgetItem()
-                itemSourceFile = QtWidgets.QTableWidgetItem()
-                itemFolder.setText(srcFolder)
-                intRow = int(self.ui.listaArchivos.rowCount())
-                
-                self.ui.listaArchivos.insertRow(intRow)
-                self.ui.listaArchivos.setItem(intRow, 0, itemFolder)
-                itemSourceFile.setText(srcFile)
-                self.ui.listaArchivos.setItem(intRow, 1, itemSourceFile)
+                if self.itemAlreadyExists(srcFile) == False:
+                    
+                    intRow = int(self.ui.listaArchivos.rowCount())
+                    
+                    itemFolderSplit = QtWidgets.QTableWidgetItem()
+                    itemFolderSplit.setText(srcFolderSplit)
+                    itemFolderSplit.setToolTip(srcFolder)
+                    
+                    #itemFolder = QtWidgets.QTableWidgetItem()
+                    #itemFolder.setText(srcFolder)
+                    
+                    self.ui.listaArchivos.insertRow(intRow)
+                    self.ui.listaArchivos.setItem(intRow, 0, itemFolderSplit)
+                    # self.ui.listaArchivos.setItem(intRow, 3, itemFolder)
+                    
+                    itemSourceFile = QtWidgets.QTableWidgetItem()                    
+                    itemSourceFile.setText(srcFile)
+                    self.ui.listaArchivos.setItem(intRow, 1, itemSourceFile)
 
-                # Función de renombrado de archivos
-                strDestFile = self.renameFiles(srcFile, intSequence)
+                    # Función de renombrado de archivos
+                    strDestFile = self.renameFiles(srcFile, intSequence)
 
-                itemDestFile = QtWidgets.QTableWidgetItem()
-                itemDestFile.setText(strDestFile)
-                self.ui.listaArchivos.setItem(intRow, 2, itemDestFile)
-                
+                    itemDestFile = QtWidgets.QTableWidgetItem()
+                    itemDestFile.setText(strDestFile)
+                    self.ui.listaArchivos.setItem(intRow, 2, itemDestFile)
+                    
                 intSequence += 1
 
     def location_on_the_screen(self):
